UnMounting Animations - Starter code
Focus on UnmoutAnimations.js

The animation works just fine, but we have a problem that when we animate from 1 to 0, the Box slides of the screen but it is still there. The we are to animate from 0 back to 1, the box will slide back up from the bottom.

We will fix this in the next commit