Photo Grid Shared Element - Add a Modal View
Add the Second view that we transition to.
Rememebr, shared elemet is about having a particular view and we have a scond view that we transition to.

At this point, we can see that we have the modal view shows over the photo grid.
Note that the pointerEvents on the modal view is initially null, this means that we
the modal can actually overlay over the photo girid, but its prety much useless. ou can not interact with it bit you can interact with the screen under like still scrolling the photo grid