Photo Grid Shared Element - Hide the imageDetail textsContent
Rememeber, that from the last commit, we had the modalView that contains the imageContent on top of the photoGrid but we set the pointerEvents on the modal to none. This makes it pretty much useless. So you can still interact with the photoGrid like the modalView is not there but the imageDetails like the long texts is there

So now we just slide done the texts in the modal View so that we can see our photoGrid :)