Animated Color Picker - Transition of TextInput and Ok button
Remember we had the the inputView and the actionView on top of each other. So now we are just transitioning the items in both views. This is the trick.
A single aniationValue for that drives the transition of the elemts like TextInput and Ok button from appearing and disappearing. Thanks to interpolation!

- Hiding and revealing the OK button triggered by when we click on the Color icon
- Add an opacity animatin for the textInput in the inputView